+++ OPHD/UI/Reports/WarehouseReport.cpp
@@ -458,7 +458,7 @@ void WarehouseReport::drawLeftPanel(Renderer& renderer)
 	renderer.drawText(*FONT_MED, WH_COUNT, NAS2D::Point{width() / 2 - 10 - COUNT_WIDTH, positionY() + 35}, textColor);
 	renderer.drawText(*FONT_MED, WH_CAPACITY, NAS2D::Point{width() / 2 - 10 - CAPACITY_WIDTH, positionY() + 57}, textColor);
 
-	drawBasicProgressBar(static_cast<float>(CAPACITY_BAR_POSITION_X), positionY() + 84.0f, static_cast<float>(CAPACITY_BAR_WIDTH), 20.0f, CAPACITY_PERCENT);
+	drawBasicProgressBar(CAPACITY_BAR_POSITION_X, positionY() + 84, CAPACITY_BAR_WIDTH, 20, CAPACITY_PERCENT);
 }
